Change the default location of the 'enabled_plugins' file
During prelaunch, if an explicit plugins file location has not been specified via the RABBITMQ_ENABLED_PLUGINS_FILE env var, adopt a new location of /var/lib/rabbitmq/enabled_plugins if possible, but fallback to the original location if there is no file at the new location, and the file at the old location is both readable and writable. No change is applied for Windows.